Define DbgapFile class
This is much easier than passing a dictionary of dictionaries and checking
for None at multiple stages. Now each instance of DbgapFile knows its
type, basename, full_name, and stores the re.match output.

This is a reasonable compromise behind entirely object oriented (which
would have DbgapDirectory and DbgapConsentGroupDirectory classes) and
function-based programming. Now the functions just operate on a list of
the objects, with the objects storing important information.