Start fns to make symlinks, move getting file_sets into main()
Functions to make symlinks might not be working yet (untested so far).
Moved the _get_special_file_set and _get_phenotype_file_sets calls
into main, since that isn't really part of making the symlinks. This
way, we can also run the script to check the diffs/file classification
with or without creating symlinks (with a command line argument).