Test _check_diffs exception when files are different, bug fix in fn
Had to fix the function because check_output raises an exception if the
exit code is not 0, and diff does not exit with an exit code of 0 if
the files are different. Now we catch the subprocess exception and raise
a ValueError with a custom message instead. I think this is better than
catching the subprocess.CalledProcessError because if we change the fn
to use something other than subprocess, we can still raise a ValueError.