Update DbgapDirectoryStructureTestCase setup and _make_file_set methods
We need a var_report and data_dict for each set of phenotype or special
files created. Therefore I took the creation of the file sets out of the
setUp() method and moved them to their own _make_file_set method. That way,
we can create whichever set of files that need to be tested in the method.
For example, not all methods need subject files to run their test.