save initial directory in setUp and chdir back in tearDown of TempdirTestCase
If a test changes directory into the tempdir and forgets to change back, we
sometimes got shell errors from later tests saying that we couldn't get the
current working directory (since it has been deleted in the tearDown
method). Now we change directory back to the initial directory in the
tearDown method to prevent this problem (including in cases where a test
that chdir'd fails).