uncompress works even if it is called from a different directory
The uncompress function should behave the same whether you are in the
directory when you call it, ie uncompress('.') or if you are calling
it from a different directory, ie uncompress('/another/directory').
To change this, I had to add an argument to the tar command stating
where the tar file should be unzipped. Update unit tests to explicitly
test this behavior.