additional tests: works if phs exists but not version, exception if both exist
If the directory/<phs_dir> already exists but directory/<phs_dir>/<version>
does not, this is ok: we are just adding a new version of a study
we already have a previous version for, so we test that the function
behaves as expected.

If directory/<phs_dir>/<version> already exists, then we are trying
to add a second copy of the same version, and the function should
raise an appropriate exception. Added a test for that case.