explicitly pass directory to organize into to _make_symlinks
I realized while writing some other unit tests that it would better
if _make_symlinks didn't create a specific directory name, but instead
accepted an argument to specify where the symlinks should be created.
Note that this commit breaks the overall workflow of the script, which
will be fixed next.