do not require pedigree files since they are not required
_get_special_file_set returns None if no files are found matching
the pattern. Unfortunately this has to be checked later for the
subject and sample files, but not for the pedigree files, since
those are not required by dbgap. I added some assert statements
(probably not the best long-term solution) to organize() to make
sure that the subject and sample file sets are not None.