Require out_path argument in create_final_directory
The output directory for final files was originally hardcoded in the
create_final_directory function. This was not ideal, since we don't
want the functions themselves to assume anything about the directory
structure (other than the raw/organized structure created by this
script). Therefore I made out_path a required argument in the
create_final_directory function instead of having a default_path
keyword argument. The default path is now coded in the --outpath
argument default in the if __name__ == '__main__' section of the
script.