Bugfix in create_final_directory and unit test for nonexistent out_path
create_final_directory now raises a FileNotFoundError if out_path doesn't
exist before attempting to create the phs directory (which consequently
also raises a FileNotFoundError). Therefore we can't test for a
FileNotFoundError by itself; we need to test the error message too for
the correct path. This can be done with the self.assertRaisesRegex
method.