Check that appropriate number of consent groups exist
When organizing dbGaP files, check that the number of consent
groups is correct. This check is done by reading in subject file in
using pandas, counting the number of unique, non-zero values in the
consent column, and checking that the number of phenotype files in
a set matches the number of consent groups. Because it involves
reading in the subject file, also change the tests to produce (more)
correct subject files, with actual values for the subject id and
consent column. Also require the user to specify the name of the
consent column in the subject file, if it is not the default of
CONSENT.