Add explicit check that consent codes match filenames
Instead of checking only that the correct number of phenotype files
given the number of unique consent codes, check that all the consent
codes are found in the set of phenotype files. For example, if
consent codes 1 and 2 exists in the subject file, the set of
phenotype files must have .c1. and .c2. in their names.