Only remove '0' consent value if it exists
Some studies do not have any subjects with a consent value of 0,
which caused the _check_consent_groups function to fail with a
KeyError exception. Check that the '0' value exists before trying
to remove it. Add a test that explicitly tests this behavior. Update
the previous method to generate test data for the subject file to
not include a 0 consent value.