Do not use the first column for rownames if trailing tabs
Subject files that have trailing tabs after their data values ended
up having the first column used as the row names, which threw off
indexing and set the consent column to a different column. Fix the
problem by forcing pandas to read the file in without using the
first column as rownames with the index_col=False keyword. (Note
how much easier this was in pandas compared to R!) Add a test for
this behavior.