Fix bug with reading in subject file if ids contain hash
Some studies have apparently used the # in their subjectIDs, but
dbGaP uses the # as comment character in the header. Modify the
code that reads in the subject file to assume that a header line
starting with a hash is a comment and that a hash character not at
the beginning of the line is actually data. Add some tests to check
this behavior.