Use the third variable in Subject file for consent by default
Instead of requiring the user to pass the consent variable, assume
that the third variable in the dbGaP Subject file is for consent
in the _check_consent_groups function. Change the default argument
for the consent_variable in other functions. Add some tests to check
this and modify other tests so that they don't pass the consent
variable name by default.